Start refactoring metadata handling
